+++ tests/utils_test.lua
@@ -1,1 +1,2 @@
+require("balm/tests/u/id_generator_test")
 require("balm/tests/u/string_buffer_test")