+++ utils.lua
@@ -1,6 +1,11 @@
 return {
+  bit = require("balm/u/bit"),
   dir8 = require("balm/u/dir8"),
+  IDGenerator = require("balm/u/id_generator"),
   NaiveDatetime = require("balm/u/naive_datetime"),
+  Properties = require("balm/u/properties"),
+  Quadmap = require("balm/u/quadmap"),
+  RecordTable = require("balm/u/record_table"),
   Scheduler = require("balm/u/scheduler"),
   StringBuffer = require("balm/u/string_buffer"),
   WindowSkin = require("balm/u/window_skin"),