+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2026.5.14
+
+* Ported `IDGenerator` from project
+* Ported `RecordTable` from project
+* Ported `Properties` from project
+
 # 2026.5.9
 
 * Rewrote `balm.m.value.inspect/1`